pyproject.toml
[tool.poetry]
name = "gptravel"
version = "0.1.0"
description = "Project for travel plan creation using GPT model"
authors = ["Stefano Polo <sp.stefano.polo@gmail.com>, Roberto Corti <robertocorti.job@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python =  ">=3.9,<3.9.7 || >3.9.7,<=3.11"
geopy = "^2.3.0"
poetry-dotenv-plugin = "^0.1.0"
openai = ">=1.40.0,<2.0.0"
python-dotenv = "^1.0.0"
regex = "^2022.10.31"
requests = "^2.28.2"
numpy = "^1.24.2"
python-tsp = "^0.3.1"
streamlit = "^1.22.0"
folium = "^0.14.0"
streamlit-folium = "^0.11.1"
spacy = "^3.5.3"
wikipedia-api = "^0.6.0"
poetry = "1.8.3"

[tool.poetry.dependencies.en_core_web_md]
url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0.tar.gz"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pylint = "^2.17.0"
mypy = "^1.1.1"
types-requests = "^2.28.11.16"
autoflake = "^2.1.1"
bandit = "^1.7.5"
pre-commit = "^4.0.1"
isort = "^5.13.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.2"
pytest-coverage = "^0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
